requirements.txt
streamlit
PyPDF2
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
rank-bm25
pdf2image
googletrans
python-dotenv
numpy
pandas
plotly
Pillow
scikit-learn
